Create grid_generater
For this I used the base code and lines 20,21 to ask for n value. Next, drawgrid generated the grid by taking x and y  as the same so just i and the size in between formulates for the spacing. finally in the while loop which is the main algorithm to keep running the code, drawgrid func takes the width of screen as w, n (input) as rows and screen (defined in beginning) as area and it generated the grid as n is entered